thay-tu-online/requirements.txt
# --- Framework & Server ---
flask
python-dotenv
uvicorn
fastapi
python-multipart

# --- Google AI & Agent Kit (Quan trọng) ---
# Đây là gói chính thức chứa google.adk, không cần git clone
google-adk>=0.4.0
google-genai
duckduckgo-search

# --- Telemetry & Utilities ---
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
requests
httpx
pydantic>=2.0